# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Translation Service",
        "version": "1.0"
    },
    "paths": {
        "/api/export": {
            "get": {
                "tags": [
                    "Translations"
                ],
                "summary": "Export translations as JSON for a specific locale",
                "description": "Returns all translations for the given locale in JSON format.",
                "operationId": "exportTranslations",
                "parameters": [
                    {
                        "name": "locale",
                        "in": "query",
                        "description": "The locale of the translations to export (e.g., 'en', 'fr').",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "en"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response with translations in JSON format",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                },
                                "example": {
                                    "welcome_message": "Welcome",
                                    "error_message": "An error occurred"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Locale not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                },
                                "example": {
                                    "message": "Locale not found"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                },
                                "example": {
                                    "message": "The given data was invalid.",
                                    "errors": {
                                        "locale": [
                                            "The locale field is required."
                                        ]
                                    }
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}